Added migration for project, merged with other scripts
